refactor: Player, Dealer 중복 코드 제거
- Participant 추상 클래스 생성
- Participant에서 HitAction 인터페이스를 구현하도록 수정
