+++ src/main/java/model/participant/Dealer.java
@@ -3,20 +3,17 @@ package model.participant;
 import java.util.List;
 import model.card.Card;
 import model.card.Cards;
-import model.game.HitAction;
 
-public class Dealer implements HitAction {
+public class Dealer extends Participant {
 
     private static final int HIT_CONDITION = 17;
-
-    private Cards cards;
 
     public Dealer() {
         this(new Cards(List.of()));
     }
 
     public Dealer(Cards cards) {
-        this.cards = cards;
+        super(cards);
     }
 
     @Override
@@ -25,20 +22,7 @@ public class Dealer implements HitAction {
         return totalNumbers < HIT_CONDITION;
     }
 
-    @Override
-    public void hitCard(Card card) {
-        cards = cards.add(card);
-    }
-
     public Card getFirstCard() {
         return cards.getCards().get(0);
-    }
-
-    public int cardsSize() {
-        return cards.size();
-    }
-
-    public Cards getCards() {
-        return cards;
     }
 }
